pyproject.toml
[project]
name = "repo-analyzer"
version = "0.1.0"
description = "Work for MLSC Coherence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
